+++ test/unit/gen_class/unit_test_run3.sh
@@ -39,6 +39,7 @@ echo "Unit test: gen_class module"
 /usr/bin/python3 ./test/unit/gen_class/daemon_init.py
 /usr/bin/python3 ./test/unit/gen_class/daemon_restart.py
 /usr/bin/python3 ./test/unit/gen_class/daemon_start.py
+/usr/bin/python3 ./test/unit/gen_class/daemon2_init.py
 /usr/bin/python3 ./test/unit/gen_class/logfile_filter_ignore.py
 /usr/bin/python3 ./test/unit/gen_class/logfile_filter_keyword.py
 /usr/bin/python3 ./test/unit/gen_class/logfile_filter_regex.py